A simple implementation of a Linked List
This .cpp file is considered to be a simple implementation because the nodes only store values of type Int. The most useful methods are: insertion (methods allow insertion at end, start, and before or after a particular node), deletion, and a method to print values in the list, in order.